+++ eyeball/CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Allow non-`Send` and / or non-`'static` values
+
 # 0.1.2
 
 - Implement `Default` for `Observable`